+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.0
 - Created the `Outcome` class under roulette module
+- Create the `Bin` class as an extended collection data structure of the frozenset.
 
 # 0.1.0
 - Unit test using `unittest` class